+++ lib/CalDAV/CalendarRoot.php
@@ -1,0 +1,18 @@
+<?php
+
+namespace Sabre\CalDAV;
+
+/**
+ * This class exists to provide forward-compatiblity with sabre/dav 2.1.
+ *
+ * You are encouraged to use CalendarRoot instead of CalendarRootNode, as the
+ * latter will be removed in a future version of sabre/dav.
+ *
+ * @copyright Copyright (C) 2007-2014 fruux GmbH (https://fruux.com/).
+ * @author Evert Pot (http://evertpot.com/)
+ * @license http://sabre.io/license/ Modified BSD License
+ */
+class CalendarRoot extends CalendarRootNode {
+
+}
+