+++ lib/CalDAV/CalendarRootNode.php
@@ -13,6 +13,8 @@ use Sabre\DAVACL\PrincipalBackend;
  * @copyright Copyright (C) 2007-2014 fruux GmbH (https://fruux.com/).
  * @author Evert Pot (http://evertpot.com/)
  * @license http://sabre.io/license/ Modified BSD License
+ * @deprecated This class will be removed in a future version of sabredav.
+ *   Use CalendarRoot instead of this class.
  */
 class CalendarRootNode extends \Sabre\DAVACL\AbstractPrincipalCollection {